+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.9 branch
 ### 1.9.5
 * the `Event`, `EventDispatcher`, `EventDispatcherTrait` and `EventList` classes are deprecated and will be removed in
-  a future release. Use instead the `symfony/event-dispatcher` component.
+  a future release. Use instead the `symfony/event-dispatcher` component;
+* the `EventAssertTrait` trait is deprecated and will be removed in a future release.
 
 ### 1.9.4
 * `TestTrait::assertArrayKeysEqual()` is deprecated and will be removed in a future release;